backend/requirements.txt
# --- existing ---
pandas==2.2.2
pytest==8.3.5

# --- API & server ---
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20

# --- LLM (free) ---
google-generativeai==0.8.3

# --- RAG ---
sentence-transformers==3.4.1
faiss-cpu==1.10.0

# --- utilities ---
python-dotenv==1.1.0
pydantic==2.11.3
